+++ NEWS.md
@@ -1,3 +1,17 @@
+# RCDT 1.1.0 (2022-04-06)
+
+* New example to show that the triangulation depends on the order of the points.
+
+* Now the output of `delaunay` contains a 'rgl' mesh object.
+
+* Now the `plotDelaunay` function differentiates the border edges and the 
+constraint edges.
+
+* New feature: elevated Delaunay triangulation (or 2.5D Delaunay triangulation).
+
+* More examples in the README file.
+
+
 # RCDT 1.0.0 (2022-03-06)
 
-* First release.
+First release.